split up large rest handling file
